requirements.txt
# Core dependencies for the coding agent benchmark suite
# No external dependencies required for basic functionality
# Built-in Python modules are used: pathlib, tempfile, shutil, json, argparse

# LLM Integration Dependencies:
requests>=2.31.0  # For OpenRouter API calls
python-dotenv>=1.0.0  # For loading environment variables from .env files
dataclasses  # Built-in in Python 3.7+

# Optional dependencies for future enhancements:
# openai>=1.0.0  # For OpenAI integration
# anthropic>=0.7.0  # For Claude integration
# pytest>=7.0.0  # For testing
# black>=23.0.0  # For code formatting
# mypy>=1.0.0  # For type checking 